requirements.txt
tldextract==4.0.0
qrcode[pil]==7.4.2
pillow==10.3.0
flask==2.0.3